Smarter board placement when resizing window.
This generally works nicely if the entire board is in view.  However,
it has the slightly weird effect of shifting the board in the opposite
direction to the mouse when it is partly off screen.  e.g. move the
board partly off the left of the window, and then resize the window
from the right edge.